+++ modules/face/insightface.py
@@ -14,7 +14,7 @@ def get_app(mp_name, threshold=0.5, resolution=640):
     if not installed('insightface', reload=False, quiet=True):
         install('insightface', 'insightface', ignore=False)
         install('albumentations==1.4.3', 'albumentations', ignore=False, reinstall=True)
-        install('pydantic==1.10.15', 'pydantic', ignore=False, reinstall=True)
+        install('pydantic==1.10.21', 'pydantic', ignore=False, reinstall=True)
     if not installed('ip_adapter', reload=False, quiet=True):
         install('git+https://github.com/tencent-ailab/IP-Adapter.git', 'ip_adapter', ignore=False)